remove icons symlink (bsc#1087224)
The icons symlink is in the way of the SLE migration. We can only turn a
symlink into a directory within the same package. The Oxygen subpackage
is obsoleted by this change as it's only purpose was to play tricks with
the symlink. Note that the oxygen icons cannot be split out into a
subpackage even in the future either due to those symlink tricks.