Restore the QGroupBox top margin (bsc#1183306)
It was changed in 171a0586fe964d from 2.5ex to 10px, which is causing a
vertical cut-off to QGroupBox titles (bsc#1183306).

This is because those titles have their origin in the left top margin
but the fixed top margin (10px) could be smaller than the font size. So,
it is better to use a relative unit based on the font-size to avoid
those cuts off.